requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
sqlalchemy>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
ddgs>=0.5.0
wikipedia-api>=0.6.0
yfinance>=0.2.38
plotly>=5.18.0
streamlit-folium>=0.20.0
folium>=0.16.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
tqdm>=4.66.0
tenacity>=8.2.0
